refactor(research): extract EPUB XHTML renderer
